+++ src/components/depth-chart/display-objects/horizontal-axis.ts
@@ -2,12 +2,8 @@ import { ScaleLinear } from "d3-scale";
 
 import { Container } from "../../../renderer/display";
 import { Text } from "../../../renderer/text";
-import { AXIS_HEIGHT } from "..";
+import { AXIS_HEIGHT, FONT_SIZE } from "../depth-chart";
 
-export const volumeFormatter = new Intl.NumberFormat("en-gb", {
-  maximumFractionDigits: 0,
-  minimumFractionDigits: 0,
-});
 export class HorizontalAxis extends Container {
   private nodeByKeyValue = new Map<string, Text>();
 
@@ -18,44 +14,47 @@ export class HorizontalAxis extends Container {
   public update(
     scale: ScaleLinear<number, number>,
     width: number,
-    height: number
+    height: number,
+    resolution: number = 1
   ) {
-    const ticks = scale.ticks(width / 100);
+    const numTicks = width / resolution / 200;
+    const ticks = scale.ticks(numTicks);
+    const tickFormat = scale.tickFormat(numTicks);
 
     const enter = ticks.filter(
-      (tick) => !this.nodeByKeyValue.has(volumeFormatter.format(tick))
+      (tick) => !this.nodeByKeyValue.has(tickFormat(tick))
     );
 
     const update = ticks.filter((tick) =>
-      this.nodeByKeyValue.has(volumeFormatter.format(tick))
+      this.nodeByKeyValue.has(tickFormat(tick))
     );
 
     const exit = [...this.nodeByKeyValue.keys()].filter(
-      (node) => !(ticks.map(volumeFormatter.format).indexOf(node) !== -1)
+      (node) => !(ticks.map(tickFormat).indexOf(node) !== -1)
     );
 
     for (const node of enter) {
-      const text = new Text(volumeFormatter.format(node), {
+      const text = new Text(tickFormat(node), {
         fill: 0xa1a1a1,
         fontFamily: "monospace",
-        fontSize: 12,
+        fontSize: FONT_SIZE,
       });
 
       text.x = scale(node);
-      text.y = height - AXIS_HEIGHT + 3;
-      text.anchor.set(0.5, 0);
+      text.y = height - (resolution * AXIS_HEIGHT) / 2;
+      text.anchor.set(0.5, 0.5);
 
       text.updateText(); // TODO: Should not need to call this
 
-      this.nodeByKeyValue.set(volumeFormatter.format(node), text);
+      this.nodeByKeyValue.set(tickFormat(node), text);
       this.addChild(text);
     }
 
     for (const node of update) {
-      const text = this.nodeByKeyValue.get(volumeFormatter.format(node))!;
+      const text = this.nodeByKeyValue.get(tickFormat(node))!;
 
       text.x = scale(node);
-      text.y = height - AXIS_HEIGHT + 3;
+      text.y = height - (resolution * AXIS_HEIGHT) / 2;
     }
 
     for (const node of exit) {